api/pyproject.toml
[tool.poetry]
name = "kamesize-api"
version = "0.1.0"
description = "an api to help on image resize"
authors = ["lucaslima18 <lucas.ala1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
coloredlogs = "^15.0.1"
logger = "^1.4"
fastapi = "^0.111.1"
pillow = "^10.4.0"
sqlmodel = "^0.0.21"
alembic = "^1.13.2"
pymysql = "^1.1.1"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
